# openapi.yaml
openapi: 3.1.0
info:
  title: Batchwise API
  version: 1.0.0
  description: Multi-tenant brewery management API.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.batchwise.example
    description: Production (configurable)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Member:
      type: object
      required: [id, email, display_name, role, is_active, created_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        role: { type: string, enum: [owner, manager, brewer, sales, viewer] }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
    MemberList:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Member" } }
    PatchMemberRequest:
      type: object
      properties:
        role: { type: string, enum: [owner, manager, brewer, sales, viewer] }
        is_active: { type: boolean, description: Deactivating signs the member out and revokes their refresh tokens. }
...
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
        request_id: { type: string }
...
  responses:
    BadRequest:
      description: Validation failure or malformed JSON
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: Missing or invalid token
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    UnprocessableEntity:
      description: Business rule violation
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: Feature not enabled for this tenant
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
paths:
...
  /api/v1/members:
    get:
      operationId: listMembers
      summary: List the tenant's members (Owners and Managers)
      tags: [members]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/MemberList" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/members/{id}:
    patch:
      operationId: patchMember
      summary: Change a member's role or active flag
      description: Owners manage every member; Managers only Brewer, Sales and Viewer members, and only to one of those roles. A tenant must keep an active Owner (422 `last_owner`). Changes are audited and apply from the member's next request.
      tags: [members]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/PatchMemberRequest" } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Member" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
...
security:
  - bearerAuth: []